eliminati metodi superflui, adeguata data class alle modifiche di room

+++ app/src/main/java/com/example/brockapp/data/TransitionAverage.kt
@@ -5,5 +5,5 @@ data class TransitionAverage(
     val longitude: Double,
     val latitude: Double,
     val averageTime: String,
-    val count: Long
+    val count: Int
 )
